requirements.txt
# Deep Learning Framework (TensorFlow includes Keras)
tensorflow==2.18.0

# Scientific Computing
numpy==1.26.0
pandas==2.1.4
scipy==1.11.4

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Geospatial Processing
netCDF4==1.6.5
xarray==2023.12.0
cartopy==0.22.0
geopandas==0.14.1
shapely==2.0.2

# Web Framework for API
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.1

# Web Server
gunicorn==21.2.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Data Processing
scikit-learn==1.3.2
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring and Logging
loguru==0.7.2

# API Documentation
flask-swagger-ui==4.11.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
schedule==1.2.0
